+++ CHANGES.md
@@ -5,6 +5,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [1.0.1] - 2015-12-07
 ### Added
  * CHANGES.md file
+ * Migration and database creation scripts
+ * File for base class for tests
+ * Tests for manage scripts
 ### Changed
  * Renamed some of the configuration keys